requirements/common.txt
django==3.0.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
django-pydenticon==0.2
Pillow==8.0.1
